Chapter 25 done

+++ Chapter 22/exampleApp/src/app/core/table.component.ts
@@ -1,8 +1,8 @@
 import {Component, Inject} from '@angular/core';
 import {Product} from '../model/product.model';
 import {Model} from '../model/repository.model';
-import {MODES, SHARED_STATE, SharedState} from './sharedState.model';
-import {Observer} from 'rxjs';
+// import {MODES, SHARED_STATE, SharedState} from './sharedState.model';
+// import {Observer} from 'rxjs';
 
 @Component({
   selector: 'paTable',
@@ -11,7 +11,7 @@ import {Observer} from 'rxjs';
 export class TableComponent {
 
   constructor(private model: Model,
-              @Inject(SHARED_STATE) private observer: Observer<SharedState>) {  }
+              /* @Inject(SHARED_STATE) private observer: Observer<SharedState> */) {  }
 
   getProduct(id: number): Product {
     return this.model.getProduct(id);
@@ -22,21 +22,24 @@ export class TableComponent {
   }
 
   deleteProduct(id: number): void {
-    this.model.deleteProduct(id);
+    let flag: boolean = confirm('Are you sure?');
+    if (flag) {
+      this.model.deleteProduct(id);
+    }
   }
 
-  editProduct(key: number) {
-    this.observer.next(new SharedState(MODES.EDIT, key));
-    // this.state.id = key;
-    // this.state.mode = MODES.EDIT;
-    // this.pass.id = key;
-  }
+  // editProduct(key: number) {
+  //   this.observer.next(new SharedState(MODES.EDIT, key));
+  //  // this.state.id = key;
+  //  // this.state.mode = MODES.EDIT;
+  //  // this.pass.id = key;
+  // }
 
-  createProduct() {
-    this.observer.next(new SharedState(MODES.CREATE));
-    // this.state.id = undefined;
-    // this.state.mode = MODES.CREATE;
-  }
+  // createProduct() {
+  //   this.observer.next(new SharedState(MODES.CREATE));
+  // // this.state.id = undefined;
+  //  // this.state.mode = MODES.CREATE;
+  // }
 }
